+++ rust/src/bvh.rs
@@ -21,13 +21,13 @@ pub enum BvhContents {
 }
 
 impl Bvh {
-    pub fn new(mut objs: Vec<Box<dyn Object>>, exposure: Range<f32>) -> Self {
+    pub fn new(mut objs: Vec<Box<dyn Object>>, exposure: Range<f64>) -> Self {
         // Note: though this BVH implementation is largely derived from Peter
         // Shirley's, it does *not* use the random axis selection and sort
         // routine, because it tended to fall into pathological cases.
 
-        fn axis_range(objs: &[Box<dyn Object>], exposure: Range<f32>, axis: Axis) -> f32 {
-            let range = objs.iter().fold(f32::MAX..f32::MIN, |range, o| {
+        fn axis_range(objs: &[Box<dyn Object>], exposure: Range<f64>, axis: Axis) -> f64 {
+            let range = objs.iter().fold(f64::MAX..f64::MIN, |range, o| {
                 let bb = o.bounding_box(exposure.clone());
                 let min = bb.min[axis].min(bb.max[axis]);
                 let max = bb.min[axis].max(bb.max[axis]);
@@ -87,8 +87,8 @@ impl Object for Bvh {
     fn hit<'o>(
         &'o self,
         ray: &Ray,
-        mut t_range: Range<f32>,
-        rng: &mut dyn FnMut() -> f32,
+        mut t_range: Range<f64>,
+        rng: &mut dyn FnMut() -> f64,
     ) -> Option<HitRecord<'o>> {
         if self.bounding_box.hit(ray, t_range.clone()) {
             match &self.contents {
@@ -121,7 +121,7 @@ impl Object for Bvh {
         }
     }
 
-    fn bounding_box(&self, _exposure: Range<f32>) -> Aabb {
+    fn bounding_box(&self, _exposure: Range<f64>) -> Aabb {
         self.bounding_box
     }
 }
